Added header for achilles_mapping and main
